@@ -14,43 +14,43 @@
         "format": "prettier --write ."
     },
     "dependencies": {
-        "@floating-ui/dom": "^1.6.11",
-        "@stellar/stellar-sdk": "^13.0.0-beta.1",
+        "@floating-ui/dom": "^1.6.12",
+        "@stellar/stellar-sdk": "13.0.0-beta.1",
         "buffer": "^6.0.3",
-        "lucide-svelte": "^0.453.0",
-        "passkey-kit": "0.9.4",
-        "svelte-persisted-store": "^0.11.0"
+        "lucide-svelte": "^0.459.0",
+        "passkey-kit": "^0.10.1",
+        "svelte-persisted-store": "^0.12.0"
     },
     "devDependencies": {
         "@skeletonlabs/skeleton": "^2.10.3",
         "@skeletonlabs/tw-plugin": "^0.4.0",
-        "@sveltejs/adapter-auto": "^3.3.0",
-        "@sveltejs/kit": "^2.7.2",
-        "@sveltejs/vite-plugin-svelte": "^3.1.2",
+        "@sveltejs/adapter-auto": "^3.3.1",
+        "@sveltejs/kit": "^2.8.1",
+        "@sveltejs/vite-plugin-svelte": "^4.0.1",
         "@tailwindcss/forms": "^0.5.9",
         "@tailwindcss/typography": "^0.5.15",
-        "@types/eslint": "^8.56.12",
-        "@types/node": "^22.7.9",
-        "@typescript-eslint/eslint-plugin": "^8.11.0",
-        "@typescript-eslint/parser": "^8.11.0",
+        "@types/eslint": "^9.6.1",
+        "@types/node": "^22.9.0",
+        "@typescript-eslint/eslint-plugin": "^8.14.0",
+        "@typescript-eslint/parser": "^8.14.0",
         "autoprefixer": "^10.4.20",
         "base64url": "^3.0.1",
         "dotenv": "^16.4.5",
-        "eslint": "^9.13.0",
+        "eslint": "^9.14.0",
         "eslint-config-prettier": "^9.1.0",
         "eslint-plugin-svelte": "^2.46.0",
-        "ye_olde_guestbook": "file:./packages/ye_olde_guestbook",
         "glob": "^11.0.0",
-        "postcss": "^8.4.47",
+        "postcss": "^8.4.49",
         "prettier": "^3.3.3",
-        "prettier-plugin-svelte": "^3.2.7",
-        "svelte": "^4.2.19",
-        "svelte-check": "^4.0.5",
-        "tailwindcss": "^3.4.14",
-        "tslib": "^2.8.0",
+        "prettier-plugin-svelte": "^3.2.8",
+        "svelte": "^5.2.0",
+        "svelte-check": "^4.0.8",
+        "tailwindcss": "^3.4.15",
+        "tslib": "^2.8.1",
         "typescript": "^5.6.3",
-        "vite": "^5.4.10",
-        "vite-plugin-tailwind-purgecss": "^0.3.3"
+        "vite": "^5.4.11",
+        "vite-plugin-tailwind-purgecss": "^0.3.3",
+        "ye_olde_guestbook": "file:packages/ye_olde_guestbook"
     },
     "pnpm": {
         "overrides": {
