@@ -1,3 +1,3 @@
-import { writable, type Writable } from "svelte/store";
+import { writable, type Writable } from 'svelte/store';
 
 export const isLoading: Writable<boolean> = writable(false);
